requirements.txt
Flask~=2.3.2
requests~=2.31.0
pywin32==306
pystray~=0.19.4
Pillow~=10.0.0
cx_freeze~=6.15.5